backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
motor==3.7.1
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
qrcode==8.2
razorpay==2.0.0
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.39.0
